Create CelebrityBrands
This is a Khan Academy project in which I was to record data from a number of celebrities of my choosing. I chose celebrities which appear the most in my personalized social media feed, and sorted them based on whether or not they had brands of their own. The celebrities brands are organized by makeup brands, clothing brands, or no brands at all. The initial table contained information about each celebrity, including their name, current age, and whether or not they have a brand. The 2 proceeding tables were for celebrities with either a makeup or clothing brand, and they included the celebrity name and current age, as well as the brand name and when it was launched.  I then singled out celebrities which were younger than 35 years of age, simply because I found it to be impressive how young many of the recorded brand founders were. The final table is the minority of the people in the initial table who do not have a brand at all, and it includes their names alone.