Create GameData
This is another Khan Academy project, where I was tasked with creating a database the way an app or game might to store data. This project was to practice the altering, updating, and deleting functions. In this project, I chose to sort the data collected through a video game I've actually played so I had accurate data. There are 4 different modes in this game, so I sorted each mode by its name, my wins, my in-game "eliminations",  matches played per mode, and the hours spent in each game mode.  I chose to use arithmetic functions to calculate the hours spent, and created another query to find my total wins, matches, and hours across the board. This was a very insightful project, as I learned I've spent too much time on a game I'm not very good at.